+++ src/utils/vue/call.ts
@@ -1,0 +1,37 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import type { AnyFunc, MaybeArray } from '@/types/modules/utils'
+
+function call(funcs: MaybeArray<() => void>): void
+
+function call<A1>(funcs: MaybeArray<(a1: A1) => void>, a1: A1): void
+
+function call<A1, A2>(
+  funcs: MaybeArray<(a1: A1, a2: A2) => void>,
+  a1: A1,
+  a2: A2,
+): void
+
+function call<A1, A2, A3>(
+  funcs: MaybeArray<(a1: A1, a2: A2, a3: A3) => void>,
+  a1: A1,
+  a2: A2,
+  a3: A3,
+): void
+
+function call<A1, A2, A3, A4>(
+  funcs: MaybeArray<(a1: A1, a2: A2, a3: A3, a4: A4) => void>,
+  a1: A1,
+  a2: A2,
+  a3: A3,
+  a4: A4,
+): void
+
+function call<A extends any[]>(funcs: AnyFunc[] | AnyFunc, ...args: A) {
+  if (Array.isArray(funcs)) {
+    funcs.forEach((func) => (call as any)(func, ...args))
+  } else {
+    return funcs(...args)
+  }
+}
+
+export { call }